Initial work on class hierarchy
